requirements.txt
streamlit>=1.29  #frontend framework for webapps
supabase>=2.0.2  #supabase client fro database operations
fastapi>=0.104.1 #backend api framework
uvicorn>=0.24.0  #ASGI server for fastapi
python-dotenv>=1.0.0 #Environment variable managementsupa